Move network/packet work from WorldManager to NetworkManager
