+++ Source/ACE.Server/Network/Managers/InboundMessageManager.cs
@@ -93,7 +93,7 @@ namespace ACE.Server.Network.Managers
             {
                 if (messageHandlerInfo.Attribute.State == session.State)
                 {
-                    WorldManager.InboundMessageQueue.EnqueueAction(new ActionEventDelegate(() =>
+                    NetworkManager.InboundMessageQueue.EnqueueAction(new ActionEventDelegate(() =>
                     {
                         // It's possible that before this work is executed by WorldManager, and after it was enqueued here, the session.Player was set to null
                         // To avoid null reference exceptions, we make sure that the player is valid before the message handler is invoked.
@@ -120,7 +120,7 @@ namespace ACE.Server.Network.Managers
 
         /// <summary>
         /// The call path for this function is as follows:
-        /// InboundMessageManager.HandleClientMessage() queues work into WorldManager.InboundMessageQueue that is run in WorldManager.UpdateWorld()
+        /// InboundMessageManager.HandleClientMessage() queues work into NetworkManager.InboundMessageQueue that is run in WorldManager.UpdateWorld()
         /// That work invokes GameActionPacket.HandleGameAction() which calls this.
         /// </summary>
         public static void HandleGameAction(GameActionType opcode, ClientMessage message, Session session)
